requirements.txt
fastapi==0.86.0
mysql_connector_repackaged==0.3.1
mysql-connector-python==8.0.31
pydantic==1.10.2
uvicorn==0.19.0
